[GDD] Fix a problem with older KOMA-script versions
The option `overfullrule` (introduced in v3.24) is a synonym for the
older option `draft` (v3.00).

The reason for the new option is, that option all options provided to
`\documentclass[...]{...}` are passed on to all loaded packages and the
option `draft` does a lot of things we don't really want.

But since we set the option outside of the documentclass options anyway
this shouldn't lead to a problem.